+++ Assets/Scripts/UI/MainMenu/MainMenuWindow.cs
@@ -4,7 +4,6 @@ using UnityEngine.UI;
 
 public class MainMenuWindow : MonoBehaviour
 {
-    [SerializeField] private Button _optionsButton;
     [SerializeField] private Button _shopButton;
     [SerializeField] private Button _inventoryButton;
     [SerializeField] private Button _startMatch;
@@ -13,28 +12,13 @@ public class MainMenuWindow : MonoBehaviour
     [SerializeField] private InventoryButtonView _inventoryButtonView;
     [SerializeField] private RectTransform _xpRect;
     [SerializeField] private RectTransform _dustRect;
-    private Sequence _optionSequence;
 
     private bool _searchMatchStart;
     private void OnEnable()
     {
-        _optionsButton.onClick.AddListener(OnClickOptionsButton);
         _shopButton.onClick.AddListener(OnClickShopButton);
         _inventoryButton.onClick.AddListener(OnClickInventoryButton);
         _startMatch.onClick.AddListener(OnClickStartMatch);
-    }
-
-    private void OnClickOptionsButton()
-    {
-        if (_optionSequence != null)
-        {
-            DOTween.Kill(_optionSequence);
-        }
-        
-        _optionSequence = DOTween.Sequence();
-        _optionSequence.Append(_optionsButton.transform.DOLocalRotate(
-            new Vector3(_optionsButton.transform.localRotation.x, _optionsButton.transform.localRotation.y, -720f), 2,
-            RotateMode.FastBeyond360));
     }
 
     private void OnClickShopButton()
@@ -62,7 +46,6 @@ public class MainMenuWindow : MonoBehaviour
     
     private void OnDestroy()
     {
-        _optionsButton.onClick.RemoveListener(OnClickOptionsButton);
         _shopButton.onClick.RemoveListener(OnClickShopButton);
         _inventoryButton.onClick.RemoveListener(OnClickInventoryButton);
         _startMatch.onClick.RemoveListener(OnClickStartMatch);
